完成權限分割

+++ app/Http/Controllers/ApplysController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
+use Gate;
 use Illuminate\Http\Request;
 use App\Models\Apply;
 use App\Http\Requests\ApplyPostRequest;
@@ -15,6 +17,9 @@ class ApplysController extends Controller
      */
     public function index()
     {
+        if(Gate::denies('show', Auth::user())){
+            return redirect()->route('home');
+        }
         $obtain = Apply::orderBy('id','DESC');
         //paginate()會將結果陣列，自動格式成他需要的樣子，而其不為JSON格式陣列，故無法成為物件陣列。get()則為一JSON格式之陣列，故可被JS的物件陣列使用。
         return view('pages.apply',['mainTitle' => '活動報名','results' => $obtain->paginate(11),'obtainArr' => $obtain->get()]);
@@ -38,6 +43,9 @@ class ApplysController extends Controller
      */
     public function store(ApplyPostRequest $request)
     {
+        if(Gate::denies('show', Auth::user())){
+            return redirect()->route('home');
+        }
         Apply::create($request->except('_token'));
         return redirect()->route('apply');
     }
@@ -73,6 +81,9 @@ class ApplysController extends Controller
      */
     public function update(ApplyPostRequest $request, $id)
     {
+        if(Gate::denies('show', Auth::user())){
+            return redirect()->route('home');
+        }
         Apply::find($id)->update($request->except('_token'));
         return redirect()->route('apply');
     }
@@ -85,6 +96,9 @@ class ApplysController extends Controller
      */
     public function destroy($id)
     {
+        if(Gate::denies('show', Auth::user())){
+            return redirect()->route('home');
+        }
         Apply::destroy($id);
         return redirect()->route('apply');
     }
